Change Validate to be a method on histogram structs
